Change gpfs perf collection to use unix socket by default
With the upgrade from GPFS 5.0.5.x to 5.1.1.X IBM changed the pmcollector
configuration to listen on a unix socket rather than a TCP port for
queries.

This update queries the collector using the unix socket first if it
exists but fails back to the old TCP port if the socket doesn't exits.
This way it should work on both GPFS 5.0.5 and 5.1.1